Extract MemberFields into member_fields.dart
