+++ lib/src/sdk/members/member_fields.dart
@@ -1,0 +1,45 @@
+enum MemberFields {
+  all,
+  id,
+  activityBlocked,
+  avatarHash,
+  avatarUrl,
+  bio,
+  bioData,
+  confirmed,
+  fullName,
+  idEnterprise,
+  idEnterprisesDeactivated,
+  idMemberReferrer,
+  idPremOrgsAdmin,
+  initials,
+  memberType,
+  nonPublic,
+  nonPublicAvailable,
+  products,
+  url,
+  username,
+  status,
+  aaEmail,
+  aaEnrolledDate,
+  aaId,
+  avatarSource,
+  email,
+  gravatarHash,
+  idBoards,
+  idOrganizations,
+  idEnterprisesAdmin,
+  limits,
+  loginTypes,
+  marketingOptIn,
+  messagesDismissed,
+  oneTimeMessagesDismissed,
+  prefs,
+  trophies,
+  uploadedAvatarHash,
+  uploadedAvatarUrl,
+  premiumFeatures,
+  isAaMastered,
+  ixUpdate,
+  idBoardsPinned,
+}
