+++ lib/src/sdk/members/member_models.dart
@@ -13,52 +13,6 @@ import '../trello_models.dart'
         OrganizationsInvited,
         OrganizationsInvitedFields;
 
-enum MemberFields {
-  all,
-  id,
-  activityBlocked,
-  avatarHash,
-  avatarUrl,
-  bio,
-  bioData,
-  confirmed,
-  fullName,
-  idEnterprise,
-  idEnterprisesDeactivated,
-  idMemberReferrer,
-  idPremOrgsAdmin,
-  initials,
-  memberType,
-  nonPublic,
-  nonPublicAvailable,
-  products,
-  url,
-  username,
-  status,
-  aaEmail,
-  aaEnrolledDate,
-  aaId,
-  avatarSource,
-  email,
-  gravatarHash,
-  idBoards,
-  idOrganizations,
-  idEnterprisesAdmin,
-  limits,
-  loginTypes,
-  marketingOptIn,
-  messagesDismissed,
-  oneTimeMessagesDismissed,
-  prefs,
-  trophies,
-  uploadedAvatarHash,
-  uploadedAvatarUrl,
-  premiumFeatures,
-  isAaMastered,
-  ixUpdate,
-  idBoardsPinned,
-}
-
 typedef MemberOrganizations = Organizations;
 typedef MemberTokens = AllOrNone;
 typedef MemberCustomStickers = AllOrNone;
