+++ lib/src/sdk/members/members.dart
@@ -1,5 +1,6 @@
 export 'member_board_filter.dart';
 export 'member_client.dart';
+export 'member_fields.dart';
 export 'member_id.dart';
 export 'member_models.dart';
 export 'trello_member.dart';
